feat: add common cell file
